requirements.txt
Flask
Flask-SQLAlchemy
Flask-Marshmallow
Marshmallow
Flask-Migrate
alembic
marshmallow-sqlalchemy